COMMERCE-6974 SF Move existing classes to correct package path
