+++ modules/apps/commerce/commerce-order-web/src/main/java/com/liferay/commerce/order/web/internal/frontend/taglib/clay/data/set/provider/CommerceNotificationDataSetDataProvider.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.commerce.order.web.internal.frontend;
+package com.liferay.commerce.order.web.internal.frontend.taglib.clay.data.set.provider;
 
 import com.liferay.commerce.frontend.model.AuthorField;
 import com.liferay.commerce.frontend.model.LabelField;
