+++ modules/apps/commerce/commerce-order-web/src/main/java/com/liferay/commerce/order/web/internal/frontend/taglib/clay/data/set/view/table/CommerceBillingAddressClayListDataSetDisplayView.java
@@ -12,10 +12,11 @@
  * details.
  */
 
-package com.liferay.commerce.order.web.internal.frontend;
+package com.liferay.commerce.order.web.internal.frontend.taglib.clay.data.set.view.table;
 
 import com.liferay.commerce.order.web.internal.frontend.constants.CommerceOrderDataSetConstants;
 import com.liferay.frontend.taglib.clay.data.set.ClayDataSetDisplayView;
+import com.liferay.frontend.taglib.clay.data.set.view.list.BaseListClayDataSetDisplayView;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -24,30 +25,20 @@ import org.osgi.service.component.annotations.Component;
  */
 @Component(
 	enabled = false, immediate = true,
-	property = "clay.data.set.display.name=" + CommerceOrderDataSetConstants.COMMERCE_DATA_SET_KEY_NOTIFICATIONS,
+	property = "clay.data.set.display.name=" + CommerceOrderDataSetConstants.COMMERCE_DATA_SET_KEY_BILLING_ADDRESSES,
 	service = ClayDataSetDisplayView.class
 )
-public class CommerceNotificationDataSetDisplayView
-	implements ClayDataSetDisplayView {
+public class CommerceBillingAddressClayListDataSetDisplayView
+	extends BaseListClayDataSetDisplayView {
 
 	@Override
-	public String getContentRenderer() {
-		return "emailsList";
+	public String getDescription() {
+		return "description";
 	}
 
 	@Override
-	public String getLabel() {
-		return "emails";
-	}
-
-	@Override
-	public String getName() {
-		return "emailsList";
-	}
-
-	@Override
-	public String getThumbnail() {
-		return "emails";
+	public String getTitle() {
+		return "title";
 	}
 
 }
