+++ modules/apps/commerce/commerce-order-web/src/main/java/com/liferay/commerce/order/web/internal/frontend/taglib/clay/data/set/view/table/CommerceNotificationDataSetDisplayView.java
@@ -12,11 +12,10 @@
  * details.
  */
 
-package com.liferay.commerce.order.web.internal.frontend;
+package com.liferay.commerce.order.web.internal.frontend.taglib.clay.data.set.view.table;
 
 import com.liferay.commerce.order.web.internal.frontend.constants.CommerceOrderDataSetConstants;
 import com.liferay.frontend.taglib.clay.data.set.ClayDataSetDisplayView;
-import com.liferay.frontend.taglib.clay.data.set.view.list.BaseListClayDataSetDisplayView;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -25,20 +24,30 @@ import org.osgi.service.component.annotations.Component;
  */
 @Component(
 	enabled = false, immediate = true,
-	property = "clay.data.set.display.name=" + CommerceOrderDataSetConstants.COMMERCE_DATA_SET_KEY_BILLING_ADDRESSES,
+	property = "clay.data.set.display.name=" + CommerceOrderDataSetConstants.COMMERCE_DATA_SET_KEY_NOTIFICATIONS,
 	service = ClayDataSetDisplayView.class
 )
-public class CommerceBillingAddressClayListDataSetDisplayView
-	extends BaseListClayDataSetDisplayView {
+public class CommerceNotificationDataSetDisplayView
+	implements ClayDataSetDisplayView {
 
 	@Override
-	public String getDescription() {
-		return "description";
+	public String getContentRenderer() {
+		return "emailsList";
 	}
 
 	@Override
-	public String getTitle() {
-		return "title";
+	public String getLabel() {
+		return "emails";
+	}
+
+	@Override
+	public String getName() {
+		return "emailsList";
+	}
+
+	@Override
+	public String getThumbnail() {
+		return "emails";
 	}
 
 }
